Unify tests across scala versions
