--- core/src/test/scala-2/com/softwaremill/diffx/test/ModifyMacroTest.scala
@@ -1,28 +1,0 @@
-package com.softwaremill.diffx.test
-
-import com.softwaremill.diffx.ModifyMacro
-import org.scalatest.flatspec.AnyFlatSpec
-import org.scalatest.matchers.should.Matchers
-
-class ModifyMacroTest extends AnyFlatSpec with Matchers {
-  it should "ignore field in nested products" in {
-    ModifyMacro.modifiedFromPath[Family, String](_.first.name) shouldBe List("first", "name")
-  }
-
-  it should "ignore fields in list of products" in {
-    ModifyMacro.modifiedFromPath[Organization, String](_.people.each.name) shouldBe List("people", "name")
-  }
-
-  it should "ignore fields in product wrapped with either" in {
-    ModifyMacro.modifiedFromPath[Either[Person, Person], String](_.eachRight.name) shouldBe List("eachRight", "name")
-    ModifyMacro.modifiedFromPath[Either[Person, Person], String](_.eachLeft.name) shouldBe List("eachLeft", "name")
-  }
-
-  it should "ignore fields in product wrapped with option" in {
-    ModifyMacro.modifiedFromPath[Option[Person], String](_.each.name) shouldBe List("name")
-  }
-
-  it should "ignore fields in map of products" in {
-    ModifyMacro.modifiedFromPath[Map[String, Person], String](_.each.name) shouldBe List("name")
-  }
-}
