+++ core/src/test/scala-3/com/softwaremill/diffx/test/DiffVersionSpecificTest.scala
@@ -1,0 +1,5 @@
+package com.softwaremill.diffx.test
+
+trait DiffVersionSpecificTest { outer: DiffTest =>
+  
+}
