+++ core/src/test/scala/com/softwaremill/diffx/test/DiffModifyIntegrationTest.scala
@@ -30,8 +30,7 @@ class DiffModifyIntegrationTest extends AnyFlatSpec with Matchers {
   }
 
   it should "allow calling ignore multiple times" in {
-    implicit val d: Diff[Person] = Diff
-      .autoDerive[Person]
+    implicit val d: Diff[Person] = Diff.autoDerive[Person]
       .ignore(_.name)
       .ignore(_.age)
     compare(p1, p2).isIdentical shouldBe true
@@ -40,8 +39,7 @@ class DiffModifyIntegrationTest extends AnyFlatSpec with Matchers {
   it should "compare lists using explicit object matcher comparator" in {
     val o1 = Organization(List(p1, p2))
     val o2 = Organization(List(p2, p1))
-    implicit val orgDiff: Diff[Organization] = Diff
-      .derived[Organization]
+    implicit val orgDiff: Diff[Organization] = Diff.autoDerive[Organization]
       .modify(_.people)
       .useMatcher(
         ObjectMatcher.list[Person].byValue(_.name)
@@ -79,8 +77,7 @@ class DiffModifyIntegrationTest extends AnyFlatSpec with Matchers {
   }
 
   it should "match map entries by values" in {
-    implicit val lookupDiff: Diff[MyLookup] = Diff
-      .derived[MyLookup]
+    implicit val lookupDiff: Diff[MyLookup] = Diff.autoDerive[MyLookup]
       .modify(_.map)
       .useMatcher(
         ObjectMatcher.map[KeyModel, String].byValue
@@ -108,8 +105,7 @@ class DiffModifyIntegrationTest extends AnyFlatSpec with Matchers {
   }
 
   it should "use overrided object matcher when comparing set" in {
-    implicit val lookupDiff: Diff[Startup] = Diff
-      .derived[Startup]
+    implicit val lookupDiff: Diff[Startup] = Diff.autoDerive[Startup]
       .modify(_.workers)
       .useMatcher(ObjectMatcher.set[Person].by(_.name))
     val p2m = p2.copy(age = 33)
