+++ core/src/test/scala/com/softwaremill/diffx/test/DiffResultTest.scala
