+++ core/src/test/scala/com/softwaremill/diffx/test/DiffSemiautoTest.scala
