+++ core/src/test/scala/com/softwaremill/diffx/test/DiffStringTest.scala
