+++ core/src/test/scala/com/softwaremill/diffx/test/ModifyMacroTest.scala
