+++ core/src/test/scala/com/softwaremill/diffx/test/examples.scala
